python_env/requirements.txt
# Core dependencies
psutil>=5.9.0
pyyaml>=6.0
python-dotenv>=1.0.0

# AWS integration
boto3>=1.28.0
botocore>=1.31.0

# Docker management
docker>=6.1.0

# Async networking
aiohttp>=3.9.0
uvloop>=0.19.0
aiodns>=3.1.0

# Monitoring
prometheus-client>=0.19.0

# HTTP client (fallback)
requests>=2.31.0

# Utilities
tenacity>=8.2.0
